fix Superfluid tests
