+++ contracts/superfluid/SuperfluidFeeManager.sol
@@ -3,6 +3,8 @@ pragma solidity ^0.8.0;
 
 import "../BaseFeeManager.sol";
 
+import "hardhat/console.sol";
+
 contract SuperfluidFeeManager is BaseFeeManager {
     uint256 public tempInterest;
 
@@ -20,10 +22,11 @@ contract SuperfluidFeeManager is BaseFeeManager {
         //     SECONDS_IN_A_DAY) /
         //     SECONDS_IN_A_YEAR /
         //     HUNDRED_PERCENT_IN_BPS;
+        platformFees += tempInterest;
 
         if (borrowAmount < platformFees) revert Errors.borrowingAmountLessThanPlatformFees();
 
-        amtToBorrower = borrowAmount - platformFees - tempInterest;
+        amtToBorrower = borrowAmount - platformFees;
     }
 
     function setTempInterest(uint256 _tempInterest) public {
@@ -32,5 +35,41 @@ contract SuperfluidFeeManager is BaseFeeManager {
 
     function deleteTempInterest() public {
         delete tempInterest;
+    }
+
+    function getDueInfo(
+        BaseStructs.CreditRecord memory _cr,
+        BaseStructs.CreditRecordStatic memory _crStatic
+    )
+        public
+        view
+        virtual
+        override
+        returns (
+            uint256 periodsPassed,
+            uint96 feesAndInterestDue,
+            uint96 totalDue,
+            uint96 unbilledPrincipal,
+            int96 totalCharges
+        )
+    {
+        if (_cr.state > BS.CreditState.Delayed) {
+            (periodsPassed, feesAndInterestDue, totalDue, unbilledPrincipal, totalCharges) = super
+                .getDueInfo(_cr, _crStatic);
+        } else if (_cr.state == BS.CreditState.Approved) {
+            periodsPassed = 1;
+            totalDue = _cr.unbilledPrincipal;
+        } else if (_cr.state == BS.CreditState.GoodStanding) {
+            periodsPassed = 1;
+            totalDue = _cr.totalDue;
+        }
+    }
+
+    function calcCorrection(
+        uint256 dueDate,
+        uint256 aprInBps,
+        uint256 amount
+    ) external view virtual override returns (uint256 correction) {
+        // reuturn 0;
     }
 }
